Eliminate *Hello* message before PUT
2.3.12 set comm remote
This command sets the ASCII string that is sent to the remote TCP client
when the TCP
port is opened, where is one or more characters up to a maximum of 32
(32
bytes). If you do not wish to use a string, use a zero (0) as the
parameter.
Default: HELLO
Example: set comm remote welcome // Set the string to welcome